backend/pyproject.toml
[tool.poetry]
authors = ["Shervin Saeedi Borna <Shervinsaeedi78@gmail.com>"]
package-mode = false
readme = "README.md"
#
[tool.poetry.dependencies]
celery = "^5.4.0"
django = "^5.1"
django-cors-headers = "^4.6.0"
django-extensions = "^3.2.3"
django-filter = "^24.2"
django-ipware = "^7.0.1"
django-nested-admin = "^4.1.1"
django-redis = "^5.4.0"
djangorestframework = "^3.15.2"
djangorestframework-camel-case = "^1.4.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
flower = "^2.0.1"
mysqlclient = "^2.2.4"
phonenumbers = "^8.13.29"
pillow = "^10.3.0"
psycopg = {extras = ["binary"], version = "^3.1.19"}
pydantic = "^2.8.2"
python = "^3.12"
python-dotenv = "^1.0.1"
redis = "^5.0.2"
requests = "^2.31.0"
tqdm = "^4.66.4"
zibal-client = "^0.3.0"

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.3"
factory-boy = "^3.3.3"
mypy = "^1.15.0"
pyright = "^1.1.395"
pytest-cov = "^5.0.0"
pytest-django = "^4.8.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ecom_core.settings"
addopts = "--cov-config=.coveragerc --cov=."
python_files = "tests.py test_*.py *_tests.py"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
exclude = ["migrations", "tests"]
strict = true

[tool.ruff]
ignore = ["E501"]
select = ["E", "F", "I", "N", "UP"]
